rename music_theory header file
